move types Recepies from CatalogPage to store models
